requirements.txt
playwright>=1.45
python-dotenv>=1.0
pydantic>=2.7
pyyaml>=6.0
pandas>=2.2
tqdm>=4.66
loguru>=0.7
tenacity>=8.4
